Moved ghost tiers to utils, simplified por fill.
