+++ utils/ghost_resources.py
